requirements.txt
altair==6.0.0
attrs==25.4.0
blinker==1.9.0
blis==0.7.9
cachetools==6.2.4
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
confection==0.0.3
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.7
emoji==2.1.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
et-xmlfile==1.1.0
feedparser==6.0.12
Flask==3.1.2
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.10.8
murmurhash==1.0.9
narwhals==2.14.0
networkx==3.4.2
nltk==3.7
numpy==1.23.4
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
pathy==0.6.2
pillow==12.0.0
plotly==6.5.0
preshed==3.0.8
proto-plus==1.27.0
protobuf==5.29.5
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==1.9.2
pydeck==0.9.1
pyodbc==4.0.34
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
referencing==0.37.0
regex==2022.9.13
requests==2.28.1
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.1.2
scipy==1.9.2
sgmllib3k==1.0.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.2
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
streamlit==1.52.1
tenacity==9.1.2
thinc==8.1.4
threadpoolctl==3.1.0
toml==0.10.2
tornado==6.5.4
tqdm==4.64.1
typer==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==1.26.12
wasabi==0.10.1
watchdog==6.0.0
Werkzeug==3.1.4
